rename context for sammensatt kontrollsak
